// src/cube/board/post/model/PostRdoModel.ts
export class PostRdoModel {
  //
  title: string = '';
  pinned: string = '';
  answered: string = '';
  companyCode: string = '';
  categoryName: string = '';
  writerName: string = '';
  searchAll: string = '';
  startDate: number = 0;
  endDate: number = 0;
  limit: number = 0;
  offset: number = 0;
}